requirements.txt
discord.py
aiohttp
supabase